requirements.txt
openai==0.27.7
langchain==0.0.220
tabulate==0.9.0
streamlit==1.22.0
pyairtable==1.5.0
python-dotenv==1.0.1